Separate NavigationStore from EphemeralStore
